fixed back to cpp

+++ src/modules/common/rawstr4.cpp
@@ -27,7 +27,6 @@
 #include <fcntl.h>
 #include <errno.h>
 #include <stdlib.h>
-#include <stdint.h>
 #include <sys/types.h>
 
 #include <utilstr.h>
@@ -114,7 +113,7 @@ void RawStr4::getIDXBufDat(long ioffset, char **buf) const
 {
 	int size;
 	char ch;
-	if ((uintptr_t)datfd > 0) {
+	if ((unsigned long)datfd > 0) {
 		datfd->seek(ioffset, SEEK_SET);
 		for (size = 0; datfd->read(&ch, 1) == 1; size++) {
 			if ((ch == '\\') || (ch == 10) || (ch == 13))
@@ -148,7 +147,7 @@ void RawStr4::getIDXBuf(long ioffset, char **buf) const
 {
 	SW_u32 offset;
 	
-	if ((uintptr_t)idxfd > 0) {
+	if ((unsigned long)idxfd > 0) {
 		idxfd->seek(ioffset, SEEK_SET);
 
 		idxfd->read(&offset, 4);
