+++ src/modules/filters/cipherfil.cpp
@@ -23,7 +23,6 @@
 
 
 #include <stdlib.h>
-#include <stdint.h>
 #include <cipherfil.h>
 #include <swcipher.h>
 #include <swbuf.h>
@@ -57,7 +56,7 @@ char CipherFilter::processText(SWBuf &text, const SWKey *key, const SWModule *mo
 			text.setSize(len + 5);
 			memcpy(text.getRawData(), cipher->getUncipheredBuf(), len);
 		}
-		else if ((uintptr_t)key == 1) {
+		else if ((unsigned long)key == 1) {
 			cipher->setUncipheredBuf(text.getRawData(), len);
 			cipher->getCipheredBuf(&len);
 			text.setSize(len + 5);
