+++ src/modules/filters/scsuutf8.cpp
@@ -30,7 +30,6 @@
  */
 
 #include <scsuutf8.h>
-#include <stdint.h>
 #include <swbuf.h>
 #ifdef _ICU_
 #include <unicode/unistr.h>
@@ -140,7 +139,7 @@ int SCSUUTF8::UTF8Output(unsigned long uchar, SWBuf* utf8Buf)
 #endif
 
 char SCSUUTF8::processText(SWBuf &text, const SWKey *key, const SWModule *module) {
-	if ((uintptr_t)key < 2)	// hack, we're en(1)/de(0)ciphering
+	if ((unsigned long)key < 2)	// hack, we're en(1)/de(0)ciphering
 		return -1;
 	
 #ifdef _ICU_
