+++ src/modules/filters/utf8arshaping.cpp
@@ -24,7 +24,6 @@
 #ifdef _ICU_
 
 #include <stdlib.h>
-#include <stdint.h>
 
 #include <utilstr.h>
 
@@ -44,7 +43,7 @@ UTF8arShaping::~UTF8arShaping() {
 char UTF8arShaping::processText(SWBuf &text, const SWKey *key, const SWModule *module)
 {
         UChar *ustr, *ustr2;
-	 if ((uintptr_t)key < 2)	// hack, we're en(1)/de(0)ciphering
+	 if ((unsigned long)key < 2)	// hack, we're en(1)/de(0)ciphering
 		return -1;
 
         int32_t len = text.length();
