+++ src/modules/filters/utf8scsu.cpp
@@ -24,7 +24,6 @@
 
 #include <utf8scsu.h>
 #include <swbuf.h>
-#include <stdint.h>
 
 SWORD_NAMESPACE_START
 
@@ -43,7 +42,7 @@ UTF8SCSU::~UTF8SCSU() {
 }
 
 char UTF8SCSU::processText(SWBuf &text, const SWKey *key, const SWModule *module) {
-	if ((uintptr_t)key < 2)	// hack, we're en(1)/de(0)ciphering
+	if ((unsigned long)key < 2)	// hack, we're en(1)/de(0)ciphering
 		return -1;
 
 	err = U_ZERO_ERROR;
